requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.7.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
GetOldTweets3==0.0.11
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lxml==4.6.2
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.6
pyquery==1.4.3
pyrsistent==0.17.3
PySocks==1.7.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
senticnet==1.6
six==1.15.0
snscrape==0.3.5.dev96+g47fbc2a
soupsieve==2.2.1
tweepy==3.10.0
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==0.16.0
